LayoutTests/html5lib/runner-expected-html5.txt

CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 2: PASS
CONSOLE MESSAGE: line 2: FOO<span>BAR</span>BAZ
Content-Type: text/plain
resources/tests1.dat: PASS

resources/tests2.dat: PASS

resources/tests3.dat: PASS

resources/tests4.dat:
3
5
6
7

Test 3 of 7 in resources/tests4.dat failed. Input:
textarea content with <em>pseudo</em> <foo>markup
Got:
| "textarea content with "
| <em>
|   "pseudo"
| " "
| <foo>
|   "markup"
Expected:
| "textarea content with <em>pseudo</em> <foo>markup"

Test 5 of 7 in resources/tests4.dat failed. Input:
</plaintext>
Got:
| 
Expected:
| "</plaintext>"

Test 6 of 7 in resources/tests4.dat failed. Input:
setting html's innerHTML
Got:
| "setting html's innerHTML"
Expected:
| <head>
| <body>
|   "setting html's innerHTML"

Test 7 of 7 in resources/tests4.dat failed. Input:
<title>setting head's innerHTML</title>
Got:
| 
Expected:
| <title>
|   "setting head's innerHTML"
resources/tests5.dat: PASS

resources/tests6.dat:
27
30
37
44
45
47

Test 27 of 51 in resources/tests6.dat failed. Input:
foo<col>
Got:
| 
Expected:
| <col>

Test 30 of 51 in resources/tests6.dat failed. Input:
</frameset><frame>
Got:
| 
Expected:
| <frame>

Test 37 of 51 in resources/tests6.dat failed. Input:
<caption><col><colgroup><tbody><tfoot><thead><tr>
Got:
| <caption>
| <colgroup>
| <tbody>
| <tfoot>
| <thead>
|   <tr>
Expected:
| <tr>

Test 44 of 51 in resources/tests6.dat failed. Input:
</table><tr>
Got:
| <tr>
Expected:
| <tbody>
|   <tr>

Test 45 of 51 in resources/tests6.dat failed. Input:
<body></body></html>
Got:
| 
Expected:
| <head>
| <body>

Test 47 of 51 in resources/tests6.dat failed. Input:
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"><html></html>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "">
| <html>
|   <head>
|   <body>
resources/tests7.dat:
24
30

Test 24 of 30 in resources/tests7.dat failed. Input:
<body>X</body></body>
Got:
| "X"
Expected:
| <head>
| <body>
|   "X"

Test 30 of 30 in resources/tests7.dat failed. Input:
<select><keygen>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
|       <option>
|         "2048 (High Grade)"
|       <option>
|         "1024 (Medium Grade)"
|       <option>
|         "512 (Low Grade)"
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
resources/tests8.dat: PASS

resources/tests9.dat: PASS

resources/tests10.dat: PASS

resources/tests11.dat: PASS

resources/tests12.dat: PASS

resources/tests14.dat: PASS

resources/tests15.dat: PASS

resources/tests16.dat: PASS

resources/webkit01.dat: PASS

resources/doctype01.dat: PASS

resources/scriptdata01.dat: PASS

resources/html5test-com.dat:
7
9
10
11
20

Test 7 of 24 in resources/html5test-com.dat failed. Input:
&lang;&rang;
Got:
| <html>
|   <head>
|   <body>
|     "〈〉"
Expected:
| <html>
|   <head>
|   <body>
|     "⟨⟩"

Test 9 of 24 in resources/html5test-com.dat failed. Input:
&ImaginaryI;
Got:
| <html>
|   <head>
|   <body>
|     "&ImaginaryI;"
Expected:
| <html>
|   <head>
|   <body>
|     "ⅈ"

Test 10 of 24 in resources/html5test-com.dat failed. Input:
&Kopf;
Got:
| <html>
|   <head>
|   <body>
|     "&Kopf;"
Expected:
| <html>
|   <head>
|   <body>
|     "𝕂"

Test 11 of 24 in resources/html5test-com.dat failed. Input:
&notinva;
Got:
| <html>
|   <head>
|   <body>
|     "&notinva;"
Expected:
| <html>
|   <head>
|   <body>
|     "∉"

Test 20 of 24 in resources/html5test-com.dat failed. Input:
<table><form><input type=hidden><input></form><div></div></table>
Got:
| <html>
|   <head>
|   <body>
|     <div>
|     <table>
|       <form>
|       <input>
|         type="hidden"
|       <input>
Expected:
| <html>
|   <head>
|   <body>
|     <input>
|     <div>
|     <table>
|       <form>
|       <input>
|         type="hidden"
resources/entities01.dat:
2
5

Test 2 of 68 in resources/entities01.dat failed. Input:
FOO&gtBAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO&gtBAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO>BAR"

Test 5 of 68 in resources/entities01.dat failed. Input:
I'm &notit; I tell you
Got:
| <html>
|   <head>
|   <body>
|     "I'm &notit; I tell you"
Expected:
| <html>
|   <head>
|   <body>
|     "I'm ¬it; I tell you"
resources/entities02.dat: PASS

resources/comments01.dat: PASS

resources/adoption01.dat:
3
11
13

Test 3 of 13 in resources/adoption01.dat failed. Input:
<a>1<button>2</a>3</button>
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "23"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "2"
|     "3"

Test 11 of 13 in resources/adoption01.dat failed. Input:
<p><b id="A"><script>document.getElementById("A").id = "B"</script></p>TEXT</b>
Got:
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="B"
|         <script>
|           "document.getElementById("A").id = "B""
|     <b>
|       id="B"
|       "TEXT"
Expected:
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="B"
|         <script>
|           "document.getElementById("A").id = "B""
|     <b>
|       id="A"
|       "TEXT"

Test 13 of 13 in resources/adoption01.dat failed. Input:
<table>A<td>B</td>C</table>
Got:
| <html>
|   <head>
|   <body>
|     "AC"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "B"
Expected:
| <html>
|   <head>
|   <body>
|     "A"
|     "C"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "B"
resources/inbody01.dat: PASS

resources/isindex.dat: PASS

resources/tables01.dat: PASS

resources/tricky01.dat:
2
6

Test 2 of 9 in resources/tricky01.dat failed. Input:
<html>
<font color=red><i>Italic and Red<p>Italic and Red </font> Just italic.</p> Italic only.</i> Plain
<p>I should not be red. <font color=red>Red. <i>Italic and red.</p>
<p>Italic and red. </i> Red.</font> I should not be red.</p>
<b>Bold <i>Bold and italic</b> Only Italic </i> Plain
Got:
| <html>
|   <head>
|   <body>
|     <font>
|       color="red"
|       <i>
|         "Italic and Red"
|     <i>
|       <p>
|         <font>
|           "Italic and Red "
|         " Just italic."
|       " Italic only."
|     " Plain
"
|     <p>
|       "I should not be red. "
|       <font>
|         color="red"
|         "Red. "
|         <i>
|           "Italic and red."
|     <font>
|       color="red"
|       <i>
|         "
"
|     <p>
|       <font>
|         <i>
|           "Italic and red. "
|         " Red."
|       " I should not be red."
|     "
"
|     <b>
|       "Bold "
|       <i>
|         "Bold and italic"
|     <i>
|       " Only Italic "
|     " Plain"
Expected:
| <html>
|   <head>
|   <body>
|     <font>
|       color="red"
|       <i>
|         "Italic and Red"
|     <i>
|       <p>
|         <font>
|           color="red"
|           "Italic and Red "
|         " Just italic."
|       " Italic only."
|     " Plain
"
|     <p>
|       "I should not be red. "
|       <font>
|         color="red"
|         "Red. "
|         <i>
|           "Italic and red."
|     <font>
|       color="red"
|       <i>
|         "
"
|     <p>
|       <font>
|         color="red"
|         <i>
|           "Italic and red. "
|         " Red."
|       " I should not be red."
|     "
"
|     <b>
|       "Bold "
|       <i>
|         "Bold and italic"
|     <i>
|       " Only Italic "
|     " Plain"

Test 6 of 9 in resources/tricky01.dat failed. Input:
<TABLE>
  <CENTER>
  <FONT>You should see a 100x100 green Image below.
  </CENTER>
<IMG alt="Image" width=100 height=100 style="background-color:green">
<tr><td>
</td>
</tr>
</table>
Got:
| <html>
|   <head>
|   <body>
|     <center>
|       "
  "
|       <font>
|         "You should see a 100x100 green Image below.
  "
|     <font>
|       <img>
|         alt="Image"
|         height="100"
|         style="background-color:green"
|         width="100"
|       "
"
|     <table>
|       "
  
"
|       <tbody>
|         <tr>
|           <td>
|             "
"
|           "
"
|         "
"
Expected:
| <html>
|   <head>
|   <body>
|     <center>
|       "
  "
|       <font>
|         "You should see a 100x100 green Image below.
  "
|     <font>
|       "
"
|       <img>
|         alt="Image"
|         height="100"
|         style="background-color: green;"
|         width="100"
|       "
"
|     <font>
|       "
"
|     <font>
|       "
"
|     <table>
|       "
  "
|       <tbody>
|         <tr>
|           <td>
|             "
"
#EOF
